refactor: group frontend components by feature domain (#140)
Reorganize the flat web/src/components/ directory into feature-based
subdirectories for better navigability and domain cohesion:

- editor/ (10 files): chapter-editor, editor-sidebar, editor-toolbar,
  editor-writing-area, editor-dialogs, ai-rewrite-sheet,
  floating-action-bar, crash-recovery-dialog, onboarding-tooltips,
  save-indicator
- drive/ (7 files): drive-banner, drive-browser-sheet, drive-files-sheet,
  drive-status-indicator, add-source-sheet, source-viewer-sheet,
  sources-panel
- project/ (9 files): delete-project-dialog, rename-project-dialog,
  duplicate-project-dialog, delete-chapter-dialog,
  disconnect-drive-dialog, project-switcher, export-menu,
  settings-menu, accounts-sheet
- layout/ (1 file): sidebar

All imports updated across pages, hooks, components, and tests.
Within-group imports use relative paths; cross-group imports use
absolute @/components/<group>/ paths.

Closes #140
