+++ web/src/app/(protected)/editor/[projectId]/page.tsx
@@ -3,14 +3,14 @@
 import { useEffect, useState, useRef } from "react";
 import { useParams, useRouter } from "next/navigation";
 import { useAuth } from "@clerk/nextjs";
-import type { ChapterData } from "@/components/sidebar";
-import type { ChapterEditorHandle } from "@/components/chapter-editor";
-import { OnboardingTooltips } from "@/components/onboarding-tooltips";
-import { CrashRecoveryDialog } from "@/components/crash-recovery-dialog";
-import { EditorSidebar } from "@/components/editor-sidebar";
-import { EditorToolbar } from "@/components/editor-toolbar";
-import { EditorWritingArea } from "@/components/editor-writing-area";
-import { EditorDialogs } from "@/components/editor-dialogs";
+import type { ChapterData } from "@/components/layout/sidebar";
+import type { ChapterEditorHandle } from "@/components/editor/chapter-editor";
+import { OnboardingTooltips } from "@/components/editor/onboarding-tooltips";
+import { CrashRecoveryDialog } from "@/components/editor/crash-recovery-dialog";
+import { EditorSidebar } from "@/components/editor/editor-sidebar";
+import { EditorToolbar } from "@/components/editor/editor-toolbar";
+import { EditorWritingArea } from "@/components/editor/editor-writing-area";
+import { EditorDialogs } from "@/components/editor/editor-dialogs";
 import { useDriveAccounts } from "@/hooks/use-drive-accounts";
 import { useDriveFiles } from "@/hooks/use-drive-files";
 import { useChapterSources } from "@/hooks/use-chapter-sources";
