+++ web/src/components/editor/crash-recovery-dialog.tsx
