+++ web/src/components/editor/floating-action-bar.tsx
