+++ web/src/components/editor/onboarding-tooltips.tsx
