+++ web/src/components/project/delete-chapter-dialog.tsx
