+++ web/src/components/project/delete-project-dialog.tsx
