+++ web/src/components/project/duplicate-project-dialog.tsx
